✨ feat     : 스터디 신청서 기능 추가
