+++ apps/category/views.py
@@ -1,9 +1,10 @@
+from unicodedata import category
+
 from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
 from rest_framework import generics
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from unicodedata import category
 
 from apps.category.models import Category
 from apps.category.serializers import CategorySerializer
